services/notification-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.9.0
pydantic-settings==2.5.0
redis==5.0.0
aioredis==2.0.1
aio-pika==9.3.0
aiohttp==3.9.0
python-telegram-bot==20.7
sendgrid==6.11.0
python-json-logger==2.0.7
cryptography==41.0.7
PyJWT==2.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
